refactor: major refactor + deletion of old code
