+++ src/lib/constants/placeholder-songsterr-url.constant.ts
@@ -1,0 +1,11 @@
+const urlPrefix = 'https://www.songsterr.com/a/wsa';
+
+// some dope songs
+export const placeholderSongUrls = [
+  `${urlPrefix}/structures-hydroplaning-tab-s88503`,
+  `${urlPrefix}/chon-fluffy-tab-s399673`,
+  `${urlPrefix}/sentinels-albatross-tab-s529860`,
+  `${urlPrefix}/vildhjarta-dagger-tab-s87817`,
+  `${urlPrefix}/humanitys-last-breath-human-swarm-tab-s390519`,
+  `${urlPrefix}/tesseract-nocturne-tab-s383709`
+];
