--- src/lib/constants/placeholderSongUrls.ts
@@ -1,11 +1,0 @@
-const baseUrl = 'https://www.songsterr.com/a/wsa';
-
-// some dope songs
-export const placeholderSongUrls = [
-  `${baseUrl}/structures-hydroplaning-tab-s88503`,
-  `${baseUrl}/chon-fluffy-tab-s399673`,
-  `${baseUrl}/sentinels-albatross-tab-s529860`,
-  `${baseUrl}/vildhjarta-dagger-tab-s87817`,
-  `${baseUrl}/humanitys-last-breath-human-swarm-tab-s390519`,
-  `${baseUrl}/tesseract-nocturne-tab-s383709`
-];
