--- src/lib/mocks.ts
@@ -1,9 +1,0 @@
-export function createGetMockSearchResultResponse(): IPartialSearchResult {
-  return {
-    songId: 0,
-    artistId: 0,
-    title: 'You are the best',
-    artist: 'Thanks for using this app',
-    source: ''
-  };
-}
