+++ src/lib/mocks/search-result.mock.ts
@@ -1,0 +1,11 @@
+function createGetMockSearchResultResponse(): IPartialSearchResult {
+  return {
+    songId: 0,
+    artistId: 0,
+    title: 'You are the best',
+    artist: 'Thanks for using this app',
+    source: ''
+  };
+}
+
+export const mockSearchResult = createGetMockSearchResultResponse();
