--- src/lib/sharedCssClasses.ts
@@ -1,8 +1,0 @@
-const sharedBtnStyle =
-  'w-fit px-2 py-1 font-semibold p-2 rounded-lg shadow-md transition duration-75 cursor-pointer';
-
-export const cssClasses = {
-  textInput:
-    'bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500 mb-2 text-center',
-  downloadBtn: sharedBtnStyle + ' bg-red-500 hover:bg-red-400 text-white'
-};
