+++ src/ui/ByLink/Form.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { enhance } from '$app/forms';
   import { toast } from '@zerodevx/svelte-toast';
-  import { selectedSongToDownload } from '../../stores/selectedSong';
+  import { selectedSongToDownload } from '../../stores/selected-song.store';
   import {
     setValidationMessage,
     clearValidationMessage
@@ -9,9 +9,9 @@
   import SelectedForm from '../withSelectedSong/SelectedForm.svelte';
   import { SONGSTERR_URL_REGEX_PATTERN } from '$lib/constants';
   import { isUrlFromSongsterr } from '$lib/utils/url';
-  import SelectedSongSkeleton from '../common/skeletons/SelectedSongSkeleton.svelte';
+  import SelectedSongSkeleton from '../general/SelectedSongSkeleton.svelte';
   import { sample } from 'lodash-es';
-  import { placeholderSongUrls } from '$lib/constants/placeholderSongUrls';
+  import { placeholderSongUrls } from '$lib/constants/placeholder-songsterr-url.constant';
 
   let selectedSong: ISearchResult | IPartialSearchResult | undefined;
   selectedSongToDownload.subscribe((value) => {
